+++ arxivabs/browse/util/id_patterns.py
