+++ browse/services/document/cache.py
@@ -5,7 +5,7 @@ from functools import wraps
 from typing import Optional
 
 from arxiv.base.globals import get_application_config, get_application_global
-from browse.domain.metadata import DocMetadata
+from arxivabs.browse.domain.metadata import DocMetadata
 
 
 # Formats that currently reside in the cache filesystem
