+++ tests/test_docmetadata.py
@@ -2,7 +2,7 @@
 from typing import Any, Dict
 from unittest import TestCase
 
-from browse.domain.metadata import DocMetadata
+from arxivabs.browse.domain.metadata import DocMetadata
 
 
 class DocMetadataTest(TestCase):
