+++ tests/test_inspire.py
@@ -6,7 +6,7 @@ from bs4 import BeautifulSoup
 
 from tests.test_abs_parser import ABS_FILES
 from browse.services.document.metadata import AbsMetaSession
-from browse.domain.license import ASSUMED_LICENSE_URI
+from arxivabs.browse.domain.license import ASSUMED_LICENSE_URI
 from browse.services.util.external_refs_cits import include_inspire_link, \
     get_orig_publish_date, INSPIRE_REF_CIT_CATEGORIES
 from hamcrest import *
