+++ tests/test_license.py
@@ -1,10 +1,10 @@
 from unittest import TestCase
-from browse.domain.license import ASSUMED_LICENSE_URI
+from arxivabs.browse.domain.license import ASSUMED_LICENSE_URI
 
 
 class TestLicense(TestCase):
     def test_new_license(self):
-        from browse.domain.license import License
+        from arxivabs.browse.domain.license import License
 
         r_uri = 'http://bogus/license/uri'
         lic = License(r_uri)
